compleed the project

+++ src/components/Demo.jsx
@@ -1,6 +1,6 @@
-import { useState, useEffect } from "react";
-import { copy, linkIcon, loader, tick } from "../assets";
+import React, { useState, useEffect } from "react";
 
+import { copy, linkIcon, loader, tick } from "../assets";
 import { useLazyGetSummaryQuery } from "../services/article";
 
 const Demo = () => {
@@ -8,15 +8,18 @@ const Demo = () => {
     url: "",
     summary: "",
   });
-
   const [allArticles, setAllArticles] = useState([]);
+  const [copied, setCopied] = useState("");
 
+  // RTK lazy query
   const [getSummary, { error, isFetching }] = useLazyGetSummaryQuery();
 
+  // Load data from localStorage on mount
   useEffect(() => {
     const articlesFromLocalStorage = JSON.parse(
       localStorage.getItem("articles")
     );
+
     if (articlesFromLocalStorage) {
       setAllArticles(articlesFromLocalStorage);
     }
@@ -25,21 +28,39 @@ const Demo = () => {
   const handleSubmit = async (e) => {
     e.preventDefault();
 
-    const { data } = await getSummary({ articleUrl: article.url });
+    const existingArticle = allArticles.find(
+      (item) => item.url === article.url
+    );
 
+    if (existingArticle) return setArticle(existingArticle);
+
+    const { data } = await getSummary({ articleUrl: article.url });
     if (data?.summary) {
       const newArticle = { ...article, summary: data.summary };
       const updatedAllArticles = [newArticle, ...allArticles];
+      const uniqueAllArticles = [...new Set(updatedAllArticles)];
+      // console.log(uniqueAllArticles);
 
+      // update state and local storage
       setArticle(newArticle);
       setAllArticles(updatedAllArticles);
 
-      localStorage.setItem("articles", JSON.stringify(updatedAllArticles));
-
       //this line of code clears the input field after the summary is fetched
       setArticle({ ...article, url: "", summary: data.summary });
+      localStorage.setItem("articles", JSON.stringify(updatedAllArticles));
+    }
+  };
 
-      console.log(newArticle);
+  // copy the url and toggle the icon for user feedback
+  const handleCopy = (copyUrl) => {
+    setCopied(copyUrl);
+    navigator.clipboard.writeText(copyUrl);
+    setTimeout(() => setCopied(false), 3000);
+  };
+
+  const handleKeyDown = (e) => {
+    if (e.keyCode === 13) {
+      handleSubmit(e);
     }
   };
 
@@ -53,27 +74,28 @@ const Demo = () => {
         >
           <img
             src={linkIcon}
-            alt="link_icon"
+            alt="link-icon"
             className="absolute left-0 my-2 ml-3 w-5"
           />
+
           <input
             type="url"
-            placeholder="Enter a URL"
+            placeholder="Paste the article link"
             value={article.url}
             onChange={(e) => setArticle({ ...article, url: e.target.value })}
+            onKeyDown={handleKeyDown}
             required
-            className="url_input peer"
+            className="url_input peer" // When you need to style an element based on the state of a sibling element, mark the sibling with the peer class, and use peer-* modifiers to style the target element
           />
-
           <button
             type="submit"
-            className="submit_btn peer-focus:border-gray-700 peer-focus:text-gray-700"
+            className="submit_btn peer-focus:border-gray-700 peer-focus:text-gray-700 "
           >
             <p>↵</p>
           </button>
         </form>
 
-        {/* Browse URL History */}
+        {/* Browse History */}
         <div className="flex flex-col gap-1 max-h-60 overflow-y-auto">
           {allArticles.map((item, index) => (
             <div
@@ -83,8 +105,8 @@ const Demo = () => {
             >
               <div className="copy_btn" onClick={() => handleCopy(item.url)}>
                 <img
-                  src={copy}
-                  alt="copy_icon"
+                  src={copied === item.url ? tick : copy}
+                  alt={copied === item.url ? "tick_icon" : "copy_icon"}
                   className="w-[40%] h-[40%] object-contain"
                 />
               </div>
@@ -96,7 +118,7 @@ const Demo = () => {
         </div>
       </div>
 
-      {/* Display Results */}
+      {/* Display Result */}
       <div className="my-10 max-w-full flex justify-center items-center">
         {isFetching ? (
           <img src={loader} alt="loader" className="w-20 h-20 object-contain" />
